Run next-reel attachment from worker polling
