+++ app/api/cron/attach-next-reel/route.ts
@@ -1,7 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { prisma } from "@/lib/db/client";
-import { getUserMedia, type InstagramMedia } from "@/lib/meta/client";
-import { decryptToken } from "@/lib/meta/oauth";
+import { attachPendingNextReels } from "@/lib/automation/attach-next-reel";
 
 /**
  * Binds "next reel" campaigns to a real post.
@@ -12,10 +10,6 @@ import { decryptToken } from "@/lib/meta/oauth";
  * Runs on a schedule (see vercel.json) — the campaign goes live within one
  * cron interval of the reel being posted.
  */
-
-function isReel(media: InstagramMedia): boolean {
-  return media.media_product_type === "REELS";
-}
 
 export async function GET(request: NextRequest) {
   const authHeader = request.headers.get("authorization");
@@ -28,68 +22,10 @@ export async function GET(request: NextRequest) {
     );
   }
 
-  const pending = await prisma.automation.findMany({
-    where: { pendingNextReel: true },
-    include: { instagramAccount: true },
-  });
-
-  // Group by connected account so we fetch each account's media only once.
-  const byAccount = new Map<
-    string,
-    { account: (typeof pending)[number]["instagramAccount"]; automations: typeof pending }
-  >();
-  for (const automation of pending) {
-    const key = automation.instagramAccountId;
-    const entry = byAccount.get(key);
-    if (entry) entry.automations.push(automation);
-    else byAccount.set(key, { account: automation.instagramAccount, automations: [automation] });
-  }
-
-  let bound = 0;
-  let checked = 0;
-  const failures: string[] = [];
-
-  for (const { account, automations } of byAccount.values()) {
-    checked += automations.length;
-    if (!account?.accessToken) continue;
-
-    let reels: InstagramMedia[];
-    try {
-      const token = decryptToken(account.accessToken);
-      const media = await getUserMedia(token, 25);
-      reels = media
-        .filter(isReel)
-        .sort(
-          (a, b) =>
-            new Date(a.timestamp).getTime() - new Date(b.timestamp).getTime()
-        );
-    } catch (err) {
-      failures.push(account.id);
-      console.error("[attach-next-reel] media fetch failed", account.id, err);
-      continue;
-    }
-
-    for (const automation of automations) {
-      // The "next" reel = the earliest one posted after the campaign was created.
-      const nextReel = reels.find(
-        (reel) => new Date(reel.timestamp) > automation.createdAt
-      );
-      if (!nextReel) continue;
-
-      await prisma.automation.update({
-        where: { id: automation.id },
-        data: {
-          postId: nextReel.id,
-          postUrl: nextReel.permalink ?? null,
-          pendingNextReel: false,
-        },
-      });
-      bound += 1;
-    }
-  }
+  const result = await attachPendingNextReels();
 
   return NextResponse.json({
     success: true,
-    data: { checked, bound, failedAccounts: failures.length },
+    data: result,
   });
 }
