+++ worker/dm-worker.ts
@@ -1,6 +1,7 @@
 import { createDMWorker } from "@/lib/queue/dm-worker";
 import { recordWorkerHeartbeat } from "@/lib/ops/worker-health";
 import { reconcileComments } from "@/lib/polling/comment-reconciler";
+import { attachPendingNextReels } from "@/lib/automation/attach-next-reel";
 import os from "node:os";
 
 const worker = createDMWorker();
@@ -32,6 +33,10 @@ const heartbeatTimer = setInterval(() => void heartbeat(), HEARTBEAT_INTERVAL_MS
 
 async function poll() {
   try {
+    const attached = await attachPendingNextReels();
+    if (attached.bound > 0 || attached.failedAccounts > 0) {
+      console.log("[DM Worker] Next-reel attachment:", attached);
+    }
     await reconcileComments();
   } catch (error) {
     const message = error instanceof Error ? error.message : "Unknown error";
